Added in the HandMeasurement class and half-implemented it. Will perform measurements on a hand in a specific pose and save the vital measurements for use in further computations.
Added PerspectiveScaler class. Self calibrates per device and allows measurements to scale with depth.

Big modification on the ClippedBitmapView class, streamlined the interface into drawing in the class. Now just allows generic drawing of points, lines and text. This will be fed into the HandMeasurement and PerspectiveScaler classes as they are implemented.